Moved common files
