+++ scrutiny/gui/widgets/graph_signal_tree.py
@@ -27,8 +27,8 @@ from PySide6.QtCharts import QLineSeries, QAbstractSeries, QValueAxis
 from scrutiny.gui import assets
 from scrutiny.gui.core.watchable_registry import WatchableRegistry
 from scrutiny.gui.core.scrutiny_drag_data import ScrutinyDragData, WatchableListDescriptor, SingleWatchableDescriptor
-from scrutiny.gui.dashboard_components.common.watchable_tree import WatchableStandardItem, get_watchable_icon
-from scrutiny.gui.dashboard_components.common.base_tree import BaseTreeModel, BaseTreeView, SerializableItemIndexDescriptor
+from scrutiny.gui.widgets.watchable_tree import WatchableStandardItem, get_watchable_icon
+from scrutiny.gui.widgets.base_tree import BaseTreeModel, BaseTreeView, SerializableItemIndexDescriptor
 from scrutiny import tools
 
 from scrutiny.tools.typing import *
