+++ test/gui/manual/manual_test_graph_axis.py
@@ -20,7 +20,7 @@ from scrutiny.gui import assets
 from scrutiny.gui.dashboard_components.varlist.varlist_component import VarListComponent
 from scrutiny.gui.dashboard_components.watch.watch_component import WatchComponent
 from scrutiny.gui.core.watchable_registry import WatchableRegistry
-from scrutiny.gui.dashboard_components.common.graph_signal_tree import GraphSignalModel, GraphSignalTree
+from scrutiny.gui.widgets.graph_signal_tree import GraphSignalModel, GraphSignalTree
 
 from test.gui.fake_server_manager import FakeServerManager, ServerConfig
